Add files via upload
In my reverse parking system project, I used an ultrasonic sensor to detect obstacles behind a vehicle. The TRIG pin sends a 10µs pulse, and the ECHO pin stays high while the wave travels and reflects. I use Timer0 on LPC21xx to measure the time and calculate distance in cm using the speed of sound.

I then transmit this distance over the CAN bus. CAN is used because it's reliable, reduces wiring, and is the standard in automotive applications.

At the receiver end, I decode the CAN message. Based on the distance value, I activate a buzzer with different beep patterns to indicate proximity — similar to what we see in modern cars.

I’ve also included a switch that allows manual sending of remote or data frames, and I use UART for debugging purposes.

This project simulates real-world automotive communication and sensor integration.